Commented 7-Employment_migration (Leo)
Leo commented this file in order to make the file more readable. He also flags several small but confusing points in the code, please check the "Migration_RMD_questions.lyx" for references. Please re-run the file, and you can get the new versions of the html or pdf files it generates.